// src/interfaces/favorites.interface.ts
import { Album } from "./album.interface";
import { Artist } from "./artist.interface";
import { Track } from "./track.interface";

export interface Favorites {
  artists: string[];
  albums: string[];
  tracks: string[];
}

export interface FavoritesResponse {
  artists: Artist[];
  albums: Album[];
  tracks: Track[];
}